Client : Refactor useInstitutionNameMap.ts. Repositioned
