+++ client/src/components/views/seb-server/settings-navigation/components/filters/useInstitutionFilterSection.ts
@@ -1,8 +1,8 @@
 import { computed } from "vue";
 import { translate } from "@/utils/generalUtils";
-import { useInstitutionNameMap } from "@/components/views/seb-server/settings-navigation/components/SettingsTable/composables/useInstitutionNameMap";
 import { useShowInstitutionColumn } from "@/components/views/seb-server/settings-navigation/components/SettingsTable/composables/useShowInstitutionColumn";
 import type { FilterSectionDef } from "./filterTypes";
+import { useInstitutionNameMap } from "@/composables/useInstitutionNameMap.ts";
 
 export const INSTITUTION_FILTER_KEY = "institutionId";
